BATC/assembly-qc/requirements.txt
biopython>=1.83
pandas>=2.0
matplotlib>=3.8
numpy>=1.26
